+++ packages/vsce/src/trees/treeItems/CICSWebServiceTreeItem.ts
@@ -10,26 +10,26 @@
  */
 
 import { TreeItem, TreeItemCollapsibleState } from "vscode";
-import { getIconFilePathFromName } from "../../../../utils/iconUtils";
-import { CICSRegionTree } from "../../../CICSRegionTree";
+import { getIconFilePathFromName } from "../../utils/iconUtils";
+import { CICSRegionTree } from "../CICSRegionTree";
 
-export class CICSTCPIPServiceTreeItem extends TreeItem {
-  tcpips: any;
+export class CICSWebServiceTreeItem extends TreeItem {
+  webservice: any;
   parentRegion: CICSRegionTree;
   directParent: any;
 
   constructor(
-    tcpips: any,
+    webservice: any,
     parentRegion: CICSRegionTree,
     directParent: any,
-    public readonly iconPath = getIconFilePathFromName("tcp-ip-service")
+    public readonly iconPath = getIconFilePathFromName("web-service")
   ) {
-    super(`${tcpips.name}`, TreeItemCollapsibleState.None);
+    super(`${webservice.name}`, TreeItemCollapsibleState.None);
 
-    this.tcpips = tcpips;
+    this.webservice = webservice;
     this.parentRegion = parentRegion;
     this.directParent = directParent;
-    this.contextValue = `cicstcpips.${tcpips.name}`;
+    this.contextValue = `cicswebservice.${webservice.name}`;
   }
 
   public setLabel(newLabel: string) {
